+++ app/controllers/libraries_controller.rb
@@ -1,0 +1,34 @@
+class LibrariesController < ApplicationController
+  before_action :authenticate_user!
+  before_action :set_library, only: %i[show]
+
+  def index
+    @libraries = current_user.libraries.order(:name)
+  end
+
+  def show
+  end
+
+  def new
+    @library = current_user.owned_libraries.build
+  end
+
+  def create
+    @library = current_user.owned_libraries.build(library_params)
+    if @library.save
+      redirect_to @library, notice: "Biblioteca creada."
+    else
+      render :new, status: :unprocessable_entity
+    end
+  end
+
+  private
+
+  def set_library
+    @library = current_user.libraries.friendly.find(params[:id])
+  end
+
+  def library_params
+    params.expect(library: [:name, :description])
+  end
+end