+++ db/migrate/20260419204256_create_memberships.rb
@@ -1,0 +1,13 @@
+class CreateMemberships < ActiveRecord::Migration[8.0]
+  def change
+    create_table :memberships do |t|
+      t.references :user, null: false, foreign_key: true
+      t.references :library, null: false, foreign_key: true
+      t.integer :role, null: false, default: 1
+
+      t.timestamps
+    end
+
+    add_index :memberships, [:user_id, :library_id], unique: true
+  end
+end