+++ test/factories/memberships.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :membership do
+    association :user
+    association :library
+    role { :member }
+  end
+end